+++ sources/nicgc/src/mojoshader/mojoshader_opengl.cpp
@@ -73,7 +73,7 @@
 
 #  define GL_GLEXT_LEGACY 1
 #  include <GL/gl.h>
-#  include "../../../niUI/src/GL/legacy_glext.h"
+#  include "../../../niUI/src/GL/ni_glext.h"
 
 #endif
